pdf_processor/pyproject.toml
[project]
name = "pdf-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "boto3>=1.40.52",
  "celery[sqs]>=5.5.3",
  "chromadb-client>=1.1.1",
  "dotenv>=0.9.9",
  "pypdf>=6.1.1",
  "sentence-transformers>=5.1.1",
  "watchfiles>=1.1.1",
]
[tool.uv.sources]
torch = { url = "https://download.pytorch.org/whl/cpu" }
